fixing duplicates
