--- src/modules/countRecipes.js
@@ -1,3 +1,0 @@
-const countRecipes = (recipe) => recipe.length;
-
-export default countRecipes;
