+++ src/modules/listItems.js
@@ -1,17 +1,6 @@
-import comments from './displayComments.js';
+import comments from './comment.js';
+import fetchData from './fetchAllData.js';
 import { getAPI, postAPI } from './involvementAPI.js';
-import countRecipes from './countRecipes.js';
-
-// FetchData
-const fetchData = async () => {
-  const response = await fetch(
-    'https://www.themealdb.com/api/json/v1/1/filter.php?c=chicken',
-  );
-  const data = await response.json();
-  return data.meals;
-};
-
-// list Data
 
 export const listLikes = () => {
   getAPI().then((response) => {
@@ -38,11 +27,16 @@ export const listLikes = () => {
 };
 
 const addLike = () => {
-  const heart = Array.from(document.querySelectorAll('.fas'));
+  const heart = Array.from(document.querySelectorAll('.far'));
   heart.forEach((item) => {
     item.addEventListener('click', async () => {
-      await postAPI(item.id);
-      listLikes();
+      if (item.style.color !== 'red') {
+        item.classList.remove('far');
+        item.classList.add('fas');
+        item.style.color = 'red';
+        await postAPI(item.id);
+        listLikes();
+      }
     });
   });
 };
@@ -68,7 +62,7 @@ export const listItems = async () => {
     div.id = element.idMeal;
 
     const heart = document.createElement('i');
-    heart.classList.add('fas', 'fa-heart');
+    heart.classList.add('far', 'fa-heart');
     heart.id = element.idMeal;
 
     const btn = document.createElement('button');
@@ -85,10 +79,5 @@ export const listItems = async () => {
     li.appendChild(btn);
     ul.appendChild(li);
   });
-
-  const recipe = document.querySelectorAll('.recipe-item');
-  const counter = document.querySelector('.recipes-counter');
-  counter.innerText = `Recipes (${countRecipes(recipe)})`;
-
   addLike();
 };
